+++ src/main/resources/db/migration/V1__init.sql
@@ -7,7 +7,7 @@ CREATE TABLE employee (
     id            SERIAL PRIMARY KEY,
     name          VARCHAR(255)   NOT NULL,
     department_id INT,
-    salary        NUMERIC(10, 2) NOT NULL,
+    salary        FLOAT NOT NULL,
     is_manager    BOOLEAN,
     FOREIGN KEY (department_id) REFERENCES department (id) ON DELETE CASCADE
 );
